requirements.txt
altair==5.5.0
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.1
certifi==2024.7.4
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.8
filelock==3.15.4
Flask==3.1.0
fonttools==4.53.0
frozendict==2.4.6
gitdb==4.0.12
GitPython==3.1.44
html5lib==1.1
idna==3.10
imageio==2.37.0
imbalanced-learn==0.13.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.5
lazy_loader==0.4
lightgbm==4.5.0
lime==0.2.0.1
llvmlite==0.44.0
lxml==4.9.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
multitasking==0.0.11
mysql-connector-python==9.1.0
narwhals==1.23.0
networkx==3.4.2
numba==0.61.0
numpy==2.0.0
packaging==24.1
pandas==2.2.3
patsy==1.0.1
peewee==3.17.8
pillow==10.4.0
platformdirs==4.2.2
plotly==5.24.1
protobuf==5.29.3
pyarrow==19.0.0
pydeck==0.9.1
pygame==2.6.0
Pygments==2.19.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.36.1
requests==2.32.3
requests-futures==1.0.2
rich==13.9.4
rpds-py==0.22.3
scikit-image==0.25.0
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
shap==0.46.0
six==1.16.0
sklearn-compat==0.1.3
slicer==0.0.8
smmap==5.0.2
soupsieve==2.6
statsmodels==0.14.4
streamlit==1.41.1
tenacity==9.0.0
threadpoolctl==3.5.0
tifffile==2025.1.10
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.1
ucimlrepo==0.0.7
urllib3==2.3.0
virtualenv==20.26.3
watchdog==6.0.0
webencodings==0.5.1
Werkzeug==3.1.3
xgboost==2.1.3
xlrd==2.0.1
yahooquery==2.3.7
yfinance==0.2.52
